+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add missing error handling to the `/summary` route, change HTTP error responses to return JSON
 * Add `-v` option which enables debug logging
 * Start using ESLint, add `no-console` rule to enforce use of logger instances
+* Include the unit's Modbus address in the device information
 
 ## 2.3.2